requirements.txt
sendgrid
python-dotenv
Flask
gunicorn 